chatbot/pyproject.toml
[tool.poetry]
name = "wenmoon"
version = "0.1.0"
description = ""
authors = ["Madelaine Boyd <madelaine.boyd+github@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
Flask = "^2.0.3"
python-dotenv = "^0.19.2"
discord = "^1.7.3"
requests = "^2.27.1"
jsonlines = "^3.0.0"
click = "^8.0.3"
openai = "^0.14.0"
numpy = "^1.22.2"
scipy = "^1.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
